TM Comments refactor and ui enhancement + added updateDate field to the BE/DB
