+++ frontend/src/components/dashboard/team/TMComments.jsx
@@ -1,19 +1,26 @@
 import React, { useState, useEffect } from "react";
 import AxiosInstance from "../../AxiosInstance";
+import { useForm } from "react-hook-form";
 
+import { MyMultilineTextField } from "../../forms/MyMultilineTextField";
+import { MySelectField } from "../../forms/MySelectField";
+import { Box } from "@mui/material";
 import { MyModal } from "../../forms/MyModal";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import Button from "@mui/material/Button";
+import Divider from "@mui/material/Divider";
 
 import ThumbUpIcon from "@mui/icons-material/ThumbUp";
 import ThumbDownIcon from "@mui/icons-material/ThumbDown";
+import AddIcon from "@mui/icons-material/Add";
 
-export const TMCommentsList = ({ userData_id, tm_id }) => {
+export const TMComments = ({ userData_id, tm_id, onCommentAdded }) => {
   const [comment, setComment] = useState([]);
   const [loading, setLoading] = useState(true);
   const [deleteModalOpen, setDeleteModalOpen] = useState(false);
   const [commentToDeleteId, setCommentToDeleteId] = useState(null);
+  const { handleSubmit, control, reset } = useForm();
 
   const GetComments = (id) => {
     const url = `team/teammember-comment/?id=${id}`;
@@ -62,56 +69,121 @@ export const TMCommentsList = ({ userData_id, tm_id }) => {
     });
   };
 
-  return (
-    <div>
-      <h2>Comments</h2>
-      {loading ? (
-        <p>Loading data...</p>
-      ) : comment.length > 0 ? (
-        <Card>
-          <CardContent>
-            {comment.map((commentItem) => (
-              <div key={commentItem.id}>
-                {" "}
-                {/* Changed p to div */}
-                <p>{commentItem.commentContent}</p>
-                {commentItem.isPositive ? (
-                  <div style={{ color: "green" }}>
-                    <ThumbUpIcon />
-                  </div>
-                ) : (
-                  <div style={{ color: "red" }}>
-                    <ThumbDownIcon />
-                  </div>
-                )}
-                <Button
-                  variant="text"
-                  onClick={() => handleConfirmDeleteComment(commentItem.id)}
-                >
-                  Delete
-                </Button>
-              </div>
-            ))}
-          </CardContent>
-        </Card>
-      ) : (
-        <Card>
-          <CardContent>
-            <p>No comments available</p>
-          </CardContent>
-        </Card>
-      )}
+  const submission = (data) => {
+    const Positive = data.positive === "Positive";
 
-      <MyModal
-        open={deleteModalOpen}
-        handleClose={() => setDeleteModalOpen(false)}
-        title="Confirm Deletion"
-        content={"Are you sure you want to delete the comment?"}
-        actions={[
-          { label: "Yes", onClick: handleDeleteComment },
-          { label: "No", onClick: () => setDeleteModalOpen(false) },
-        ]}
-      />
-    </div>
+    AxiosInstance.post("team/teammember-comment/", {
+      isPositive: Positive,
+      commentContent: data.comment,
+      created_by: userData_id,
+      teammember: tm_id,
+      updateDate: new Date(),
+    })
+      .then(() => {
+        onCommentAdded();
+        reset();
+      })
+      .catch((error) => {
+        console.error("Error during login", error);
+      });
+  };
+
+  return (
+    <Box>
+      <h2>Comments</h2>
+      <div>
+        <Card>
+          <CardContent>
+            <div style={{ marginBottom: "20px" }}>
+              <h3>Add new comment</h3>
+              <form onSubmit={handleSubmit(submission)}>
+                <Box>
+                  <MyMultilineTextField
+                    label={"Comment"}
+                    name={"comment"}
+                    control={control}
+                  />
+                  <MySelectField
+                    options={[
+                      { label: "Positive", value: "Positive" },
+                      { label: "Negative", value: "Negative" },
+                    ]}
+                    label={"Positive or Negative?"}
+                    name={"positive"}
+                    control={control}
+                  />
+                </Box>
+                <Button
+                  type="submit"
+                  variant="outlined"
+                  startIcon={<AddIcon />}
+                  sx={{ height: "50%" }}
+                >
+                  New comment
+                </Button>
+              </form>
+            </div>
+            <Divider />
+
+            {loading ? (
+              <p>Loading data...</p>
+            ) : comment.length > 0 ? (
+              comment.map((commentItem, index) => (
+                <div key={commentItem.id}>
+                  <div
+                    style={{
+                      display: "flex",
+                      justifyContent: "space-between",
+                    }}
+                  >
+                    <div style={{ width: "90%" }}>
+                      <p>{commentItem.commentContent}</p>
+                    </div>
+                    <div
+                      style={{
+                        position: "relative",
+                        color: "rgba(29, 33, 47, 0.4)",
+                        width: "auto",
+                      }}
+                    >
+                      <p>{commentItem.updateDate.slice(0, 10)}</p>
+                    </div>
+                  </div>
+                  {commentItem.isPositive ? (
+                    <div style={{ color: "green" }}>
+                      <ThumbUpIcon />
+                    </div>
+                  ) : (
+                    <div style={{ color: "red" }}>
+                      <ThumbDownIcon />
+                    </div>
+                  )}
+                  <Button
+                    variant="text"
+                    onClick={() => handleConfirmDeleteComment(commentItem.id)}
+                  >
+                    Delete
+                  </Button>
+                  {index < comment.length - 1 && <Divider />}
+                </div>
+              ))
+            ) : (
+              <p>You didn&apos;t add any comments so far...</p>
+            )}
+          </CardContent>
+        </Card>
+
+        <MyModal
+          open={deleteModalOpen}
+          handleClose={() => setDeleteModalOpen(false)}
+          title="Confirm Deletion"
+          content={"Are you sure you want to delete the comment?"}
+          actions={[
+            { label: "Yes", onClick: handleDeleteComment },
+            { label: "No", onClick: () => setDeleteModalOpen(false) },
+          ]}
+        />
+      </div>
+    </Box>
   );
 };
